major restructure of file hierarchy

+++ src/math/math.cpp
@@ -3,12 +3,12 @@
 namespace lyra {
 
 float pyth(const glm::vec2 a, const glm::vec2 b) {
-	double x = b.x - a.x, y = b.y - a.y;
+	float x = b.x - a.x, y = b.y - a.y;
 	return sqrt(x * x + y * y);
 }
 
 float pyth3(const glm::vec3 a, const glm::vec3 b) {
-	double x = b.x - a.x, y = b.y - a.y, z = b.z - a.z;
+	float x = b.x - a.x, y = b.y - a.y, z = b.z - a.z;
 	return sqrt(x * x + y * y + z * z);
 }
